Deploy and test toy ledger
This commit implements and end-to-end flow for building a Wasm contract,
deploying it to a node and calling it.

The contract code is contained in the `./ledger` folder. The
`./tools/build-ledger-wasm` script compiles the contract code.

We also provide two commands `osc-deploy` and `osc-ping`. The first is
for deploying the ledger contract, the second for querying it.

All this is documented in the Readme.