Project registration returns ID
With this change projects are not associated with accounts anymore.
Projects will not be identified and addressed by there associated
account but by there project IDs. In particular project registration now
returns a project ID instead of accepting a project account parameter.

The previous behavior raised a lot of questions around the special
status of an account associated with a project and the account’s private
key.

* Would the private key control all the funds of the project?
* What permissions does this private key have and how do they differ
  from other project member’s permissions that are not associated with
  that private key?
* What happens when ownership is transfered? We cannot guarantee that
  the previous owner gives up the private key.
* Would the `register_account` transaction be sent from the account
  associated with a project?