Change address of development account
We change the address of the development account that has a ton of eth.
Before we used the dev account used by the parity development chain.
However this account seems to be blocked by Parity UI if used by a chain
with a different name (`oscoin-dev` in our case). To be able to use
Parity UI properly we change the address of the dev account.