Explicitly pass sender to client methods
Instead of providing the transaction sender when `Client` is constructed
we pass it when a method is called that requires a sender. (These are
the methods that submit transactions and update the ledger.)

This allows consumers to use the client even if they don’t have an
account yet, for example if they want to create an account.