Compile ledger to Wasm with no_std
Before this change the ledger code was compiled _without_ `no_std` to
the Wasm target. This did not cause problems yet since we did not use
any `std` code. However, in the future we would have run into issues if
we inadvertently used `std` code.

The reason why `no_std` was not enabled was that the
`--no-default-features` flag was not respected in
`tool/build-ledger-wasm` because of
https://github.com/rust-lang/cargo/issues/5015.