Only upload target/cargo cache once
Instead of updating the cache of `~/.cargo` and `./target` for _every_
build we only upload it once for a specific `Cargo.lock` file.

This means that in the worst case we need to recompile all artifacts
generated from the source code in the repository even if there is a
cache. (Note that we never need to recompile dependencies because if
they change `Cargo.lock` changes.) However, recompiling the project code
is fast enough, in fact it is faster than uploading the new cache every
time. Also note that even with caching we often need to recompile large
portions of the code. The new setup also prevents the cached `./target`
directory from growing indefinitely.