Separate package to build oscoin_ledger for pwasm
We create a new package `oscoin_ledger_pwasm` that we use to compile the
ledger to Parity Wasm. This package is a simple wrapper for
`oscoin_ledger`. This allows us to compile the `oscoin_ledger` with
crate type `lib`. This enables us to compile the doc tests and reuse
`oscoin_ledger` in other crates.