Remove account unlocking from contract deployment code
This commit also updates the dependency to `rust-web3` so that
it uses a version of `web3::contract::deploy::execute` - called
`sign_and_execute` - that does not rely on account unlocking.
It instead uses transaction signing and raw transaction sending,
which do not require an account to be unlocked in order to
deploy a contract.