Extend project with additonal fields (#68)
To complete the end-to-end integration with the product MVP we need more
fields to be stored. Those are metadata fields provided by the user
creating the project. For now this is an extension of the Project entity
itself until we look more closely into proper metadata storage. It's
understood as metadata as there are no guarantees on uniqueness and are
complete free-form inputs by the user.

Field changes:
(+) name
(+) description
(=) url -> img_url

* update ledger project registration
* update client project refgistration
* update tests

***
* Extend project with additional fields
* Use to_owned consistently
* Remove trace logging from dev node
* Fix e2e tests
* Reduce number of projects in list e2e test